services/requirements.txt
djangorestframework 
markdown
django-filter
python-decouple
parinya
gunicorn
pip-autoremove
opencv-python